requirements.txt
aiofiles==0.6.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
fastapi==0.63.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
python-dotenv==0.17.0
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.8
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
